Move parser package
